input file check

+++ reference_tracks/convert_gencode_knownGene_to_sorted_refgene_track.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-echo'
+echo '
 
 Downloaded Gencode knownGene table from 
 
@@ -23,16 +23,23 @@ $12     alignID : uc031tlb.1
 '
 
 #inputFile=$1
-inputFile= gencode_v32_knownGene.txt.gz
+
+inputFile="gencode_v32_knownGene.txt.gz"
+
+if [ ! -f $inputFile ]; then
+    echo $inputFile not found
+    exit 1
+fi
+
 prefix=$(echo $inputFile | sed s/.txt// | sed s/.gz//)
 bucket="gs://tgg-viewer/ref/GRCh38/gencode_v32"
 echo $prefix
 
-cat ./$inputFile  | grep -v chrom | awk -F $'\t' 'BEGIN {OFS=FS} { print $2, $4, $5, gensub(/[.][0-9]/, "", "g", $1), "0", $3, $6, $7, "0,0,255",  $8, $9, $10 }' | grep -v _fix | grep -v _random | grep -v _alt | bedtools sort -g ~/p1/ref/GRCh38/hg38.fa.fai |  bgzip > ${prefix}.bed.gz
+gzcat ./$inputFile  | grep -v chrom | awk -F $'\t' 'BEGIN {OFS=FS} { print $2, $4, $5, gensub(/[.][0-9]/, "", "g", $1), "0", $3, $6, $7, "0,0,255",  $8, $9, $10 }' | grep -v _fix | grep -v _random | grep -v _alt |  bgzip > ${prefix}.bed.gz
 
-gzcat ${prefix}.bed.gz | awk -F $'\t' 'BEGIN {OFS=FS} { print "0", $4, $1, $6, $2, $3, $7, $8, $10, $11, $12, "0", $4 }' | bgzip  > ${prefix}.sorted.txt.gz
+bedtools sort -g ~/p1/ref/GRCh38/hg38.fa.fai -i ${prefix}.bed.gz | awk -F $'\t' 'BEGIN {OFS=FS} { print "0", $4, $1, $6, $2, $3, $7, $8, $10, $11, $12, "0", $4 }' | bgzip  > ${prefix}.sorted.txt.gz
 
 tabix -s 3 -b 5 -e 6  ${prefix}.sorted.txt.gz 
 
 gsutil -m cp ${prefix}.sorted.txt.gz* $bucket
-echo Wrote ${prefix}.sorted.txt.gz and copied it to ${bucket}/${prefix}.gtf.gz
+echo Wrote ${prefix}.sorted.txt.gz and copied it to ${bucket}/${prefix}.sorted.txt.gz
